feat: Add animations
Not sure what the pausing behaviour should be yet but this is the animation framework, still a big buggy but a good start

NOTE: There are some threading issues with the fill system, I think the data is not being shared correctly.

On the simulator the display fails if the joystick is pressed while the fill system is running. To fix.